Use the two integer variant of advisory locks
This replaces use of `(bigint)` with `(int, int)` for advisory locks.
The first `int` acts as a namespace and is derived from the unique `oid`
value for the `oban_jobs` table. The `oid` is unique within a database
and even changes on repeat table definitions.

This change aims to prevent lock collision with application level
advisory lock usage and other libraries. Now there is a 1 in
2,147,483,647 chance of colliding with other locks.