Accept a job struct as the argument to perform/1 (#32)
The `perform/1` function now receives an `Oban.Job` struct as the sole
argument, calling `perform/1` again with only the `args` map if no
clause catches the struct. This allows workers to use any attribute of
the job to customize behaviour, e.g. the number of attempts or when a
job was inserted into the database.

The implementation is entirely backward compatible, provided workers are
defined with the `use` macro. Workers that implement the `Oban.Worker`
behaviour manually will need to change the signature of `perform/1` to
accept a job struct.