Remove perform/1 callback in favor of perform/2
The new `perform/2` function receives the job's args, followed by the
complete job struct. This new function signature makes it clear that the
args are always available, and the job struct is also there when it is
needed.

This is a breaking change that will require all worker modules to be
updated.

Closes #45