Guard against running migrations repeatedly
Running the same migration repeatedly is now a no-op rather than causing
errors. This was a problem in CI systems where the `up` migration would
be ran multiple times but there wasn't a value stored on the table. In
that situation the first migration would update to the current version
and the secondary migration would error out.

This puts guards in place to ensure that repeated up and down migrations
don't cause any problems (and uses property tests to ensure a high level
of test coverage).

Closes #69