Make maxage of beat pruning configurable
Previously beats were retained for one hour and the value wasn't
configurable. That meant a lot of beats could accumulate with multiple
queues, and there wasn't any inherent value in beats older than a couple
of minutes.

This makes beat pruning configurable via a `beats_maxage` setting, which
accepts a value in seconds and must be more than a minute. The default
retention period is now five minutes (a 12x reduction in retained beats
per queue).

Closes #141