Publicize Notifier module with improved interfaces
Make the `Notifier` module public to provide the foundation of
multi-node pubsub communication. This adds documentation for the module
while cleaning up the interface to make it ergonomic for use outside of
Oban itself.

Combined with telemetry this provides everything that is necessary to
subscribe to job level events within an application.

Closes #183