Add discard and snooze support for workers (#205)
* Support snoozing jobs with a snooze return tuple

This adds support returning `{:snooze, seconds}` from `perform/2` to
re-schedule a job some number of seconds in the future. This is useful
for recycling jobs that aren't ready to run yet, e.g. because of rate
limiting.

* Support using :discard to automatically discard

Add support for returning `:discard` from `perform/2` to immediately
discard a job. This is useful when a job encounters an error that won't
resolve with time, e.g. invalid arguments or a missing record.

* Document snooze and discard functionality