Initial structure for guides (#240)
Given that the current `Oban` module takes its `@moduledoc` from the
README, I think that the "Troubleshooting" section does not belong
there. So, that's why I took that out from the README and started a
"Guides" section for `ex_doc`, which I think is more appropiate in this
case, and it could provide the space for more Guides, recipes,
or How-to's in the future.