Limit default backoff calculations to 20 attempts
The default backoff is now clamped at the equivalent of 24 days. The
change addresses an issue with snoozing, which can increase a job's
attempts into the hundreds or thousands. In this situation the algorithm
calculates the backoff using a ratio of attempts to max attempts, but is
still limited to roughly 24 days.

Closes #245