fix: handle cases where the given cron step is one
The cron configuration presented on the `#oban` slack channel the following:

```elixir
{"0 0 */1 * *", Worker.MarkerScheduler}
```

That configuration produced an `ArgumentError` exception while Oban tried to
expand the parsed result by `nimble_parsec` saying that we got an `Unexpected
value 1 outside of range 1..31`.

In this case, given that the step is 1, one quick workaround is replacing the
original configuration with:

```elixir
{"0 0 * * *", Worker.MarkerScheduler}
```

And it will produce the same result, which is to run the specified job at 00:00
on every day-of-month.

But, given that the original configuration is valid, we need to support that as
well. So, here is my attempt to fix that issue.