Add meta field and cancelled state
This adds a new migration, V9, which introduces the `meta` field, a new
`cancelled` state, and a corresponding `cancelled_at` timestamp that is
now used when cancelling jobs.

The migration requires a work-around to add a new value to the
`oban_job_state` type. As stated in
https://www.postgresql.org/docs/9.5/sql-altertype.html ALTER TYPE ...
ADD can't be used in transactions. This has been changed only in
PostgreSQL 12, so we needed a workaround.