Replace local names with registry (#331)
This replaces locally registered names with Registry based names, which allows us to use arbitrary terms as oban instance names, e.g. {Oban, name: {:tenant, 1}}. This is the preparation for supporting running dynamic Oban instances which can work with dynamic Ecto repos, as discussed in #326.

In terms of function signatures, the change should be backwards compatible. However, if the client code relied on the previous name registration, it will break.

Summary of changes:

* Oban now has its own supervision tree with a registry which is used for process discovery
* Any term can be used as name. Note that name still has to be unique, i.e. it's not possible to start two instances with the same name
* Configuration is also managed in the registry
* To improve confidence, integration tests are adapted to use a unique non-atom name
* Name is also used as the default child id in Oban.child_spec