Support dynamic repos (#347)
This introduces an `Oban.Repo` module which wraps all repo functions currently used in Oban, except for `Ecto.Repo.load`, because that one doesn't require setting prefix or dynamic repo instance pid. The module is intended to be a part of the public API, because it has to be used by plugins (otherwise they will access the wrong instance, or more likely fail because they will try to use the global instance which does not exist).

All points that called the config's `repo` directly now use the wrapper functions from `Oban.Repo`.