Rewrite cron parser for size and check performance
The nimble_parsec based generator read well, but it required us to
either bundle nimble_parsec or include the compiled output. We had
chosen to include the compiled output to minimize dependencies, and it
was over 5000 lines.

This replaces the nimble_parsec based generator with a much smaller hand
written parser along with a couple of improvements:

* Expressions are parsed into sets, simplifying membership checks
* No more special casing for the @reboot expression, instead it
  evaluates to the current time (which is unique for 7 years)