Delay supervisor init event until tree finishes
Previously, the event was fired before any children started. This led to
race conditions for handlers that awaited the init event and then
expected the presence of queues or plugins. The new approach is to emit
the event from a one-off task that boots last in the supervision tree,
which guarantees that all other children have booted.