Limit the number of jobs staged at one time
When there were a huge number of scheduled jobs the stager could time
out while updating the status to `available`. It's better to miss a few
scheduled jobs than to crash the stager—it'll stage again in another
second or so anyhow.

The limit is configurable and defaults to 5000 jobs.

Closes #477