bugfix: keep transaction behavior using dynamic repo
One of the fundamental advantages of using Oban is that it runs on top
of PostgreSQL, this setup allow us to enqueue jobs along with multiple
repo operations, ensuring that everything is committed or rolled back
atomically using transactions.

The previous behavior fails when you set a dynamic repo as part of the
Oban configuration via the `get_dynamic_repo` setting. The Oban dynamic
repo always takes precedence even when the caller is running a
transaction.

This commit/pr is an attempt to fix the previous issue.

Fixes: #442