Accept non-map terms for perform_job/3
While a worker's `new/2` callback expects a map, users may override the
function to accept any term. The `Testing.perform_job/3` function was
too strict in this regard, and prevented users from testing workers with
atypical `new/2` implementations. This loosens the API for
`perform_job/3` to make it work with any data type.

Closes #506