Add with_limit option to drain_queue
The new option makes it possible to control how many jobs are drained
totally, or, when recursion is enabled, how many jobs are processed in a
batch. This is especially helpful for workflows where many jobs may
start with intermingled dependencies. Some jobs would end up waiting for
a dependency to finish and tests could get very slow.

With this patch setting `with_limit: 1` sped up a a workflow test by
10x, cutting the total time from 2.0s to 0.2s.