Change queue supervisor strategy to :one_for_all
Queue supervisors used a `:rest_for_one` strategy, which allowed the
task supervisor to keep running when a producer crashed. That allowed
duplicate long-lived jobs to run simultaneously, which is a bug in
itself, but it could also cause `attempt > max_attempts` violations.

Switching the strategy to `:one_for_all` ensures that the task
supervisor shuts down if the producer crashes.

Closes #532