Fix error when a job exits (#588)
Sometimes Oban is unable to process certain DOWN messages. It crashes on
a line calling `Exception.format` because the "stacktrace" that we save
here is not always a stacktrace after all.

From https://hexdocs.pm/elixir/Exception.html#format/3, if kind is
{:EXIT, pid}, then it doesn't format the stacktrace.

This prevents failed processing by passing `{:EXIT, pid}` as the
processing in those circumstances.