Introduce basic lifeline plugin
When the Erlang VM is abruptly taken down (e.g. by OOM killer) or the
job does not finish quickly enough during a graceful shutdown it can be
left in the executing state forever and becomes an orphaned job.

There is no notification about this fact and the only way to learn about
it is if you stumble upon such job while inspecting the oban_jobs table.

This is a basic version without any heuristics or configuration, and it
links to the more full featured "DynamicLifeline" plugin.

Closes #603