Retain peer election schedule timing after a crash
A bug in the Peer module's "down" handler overwrote the election
scheduling interval to `0`. As soon as the leader crashed all other
peers in the cluster would start trying to acquire leadership as fast as
possible. That caused excessive database load and churn.

In addition to the interval fix, this adds a few other improvements:

* The scheduling interval is 30s rather than 15s, to space out queries
* Unknown messages, i.e. exits, are logged as warnings to help with
  debugging in the future

Closes #651