Don't fetch jobs that have reached max attempts
This adds a safeguard to the `fetch_jobs` function to prevent ever
hitting the `attempt <= max_attempts` check constraint. Hitting the
constraint causes the query to fail, which crashes the producer and
starts an infinite loop of crashes. The previous commit _should_ prevent
this situation from ocurring at the "staging" level, but to be
absolutely safe this change prevents it at the "fetching" level too.

There is a very minor performance hit from this change because the query
can no longer run as an index only scan. For systems with a modest
number of available jobs the performance impact is indistinguishable.

Closes #652