Prevent empty maps from matching non-empty maps
Due to a quirk in the jsonb containment operator (`@>`), uniquness
checks for args/meta would erroneously match an empty map (`%{}`) with
_any_ non-empty map. To prevent that mismatch, we now flip the
containment operator (`<@`) when args/meta is an empty map.

Note that there is a major performance impact when filtering by an empty
jsonb object. Postgres isn't able to use an index for any operators when
checking for an empty map and it resorts to a sequential scan instead.
That was the case before this patch, and the situation doesn't improve
when switching operators.

Closes #704