Soft replace discard with cancel return value (#730)
Discard was originally intended to mean "a job exhausted all retries."
Later, it was added as a return type and it came to mean either "stop
retrying" or "exhausted retries" ambiguously, with no clear way to
differentiate. Even later, we introduced `cancel` with a `cancelled`
state as way to stop jobs at runtime. At a semantic level, it's better
to keep discarding and cancelling distinct.

This introduces a `{:cancel, reason}` return type and soft deprecates
the use of `discard` from `perform/1` entirely. The meaning of each
action/state is now:

* `cancel`—this job was purposefully stopped from retrying, either from
  a return value or the cancel command
* `discard`—this job has exhausted all retries

Closes #703