Add prune_jobs/3 callback for engine based pruning
A dynamic approach was required for pruning queries because SQLite
doesn't support the same query features as PostgreSQL. Rather than
hard-coding an adapter into the Pruner, the engine itself gains a
`prune_jobs/3` callback. We safely fallback to `Basic.prune_jobs/3`
for engines that don't implement the callback yet.

The pruner still manages options and timed invocations, but it's up to
the engine to perform the pruning.

Closes #844