Use a CTE rather than subquery for fetch_jobs/3
The Postgres planner may choose to generate a plan that executes a
nested loop over the LIMITing subquery, causing more UPDATEs than LIMIT.
That could cause unexpected updates, including attempts > max_attempts
in some cases. The solution is to use a CTE as an "optimization fence"
that forces Postgres _not_ to optimize the query.

This is a long-standing issue that hasn't been reproducible or apparent.

More information in these links:

* https://www.2ndquadrant.com/en/blog/postgresql-ctes-are-optimization-fences/
* https://dba.stackexchange.com/questions/69471/postgres-update-limit-1/69497#69497