Match args/meta patterns in Elixir rather than PG
The containment operators, `@>` and `<@`, used for pattern matching in
tests are only available in Postgres and have some quirks. Most notably,
containment considers matching any value in a list a successful match,
which isn't intuitive or desirable.

The other issue with using a containment operator in tests is that
SQLite doesn't have those operators available and test helpers are
shared between all engines.

Closes #958, #959