Handle rollback during postgres peer election
Infrequently, the postgres peer election transaction returns `{:error,
:rollback}`. Now that return value is handled to prevent a match error.

The peer maintains its current `leader?` status on rollback—this may
cause inconsistency if the leader encounters an error and multiple
rollbacks happen in sequence. That tradeoff is acceptable because the
situation is unlikely and less of an issue than crashing the peer.

Closes #1007