Remove pruner from installation
Feel free to close this it was intentional, but I am going through these guides for the first time and I was a bit puzzled to see the pruner included in the plugin list by default. I assume it is not required, and it is covered in the next stage in the [next guide](https://hexdocs.pm/oban/preparing_for_production.html#pruning-jobs)

If it's not required, seems best to not refer to it until there is more explanation of the purpose.