Use notifier connection to deliver notifications
The Postgres notifier holds a single connection for listening and
relaying messages. However, it wasn't used to dispatch messages; that
was left to queries through the Ecto pool. Those queries are noisy and
put unnecessary load on the pool, particularly from insert
notifications. Now notifications are delivered through the notifier's
connection.