Check process and propagate notify errors
The `Notifier.notify/1` spec showed it would always return `:ok`, but
that wasn't the case when the notifier was disconnected or the process
was no longer running. Now an error tuple is returned when a notifier
process isn't running.

This situation happened most frequently during shutdown, particularly
from external usage of the Notifier like an application or the
`oban_met` package.

In addition, the errors bubble up through top level `Oban` functions
like `scale_queue/1`, `pause_queue/1`, etc. to indicate that the
operation won't actually succeed.

Closes #1060