Emit queue shutdown telemetry event with logging
A new queue shutdown event, `[:oban, :queue, :shutdown]`, is emitted by
each queue when it terminates. The event originates from the `watchman`
process, which tracks the total ellapsed time from when termination
starts to when all jobs complete or the allotted period is exhausted.

Any jobs that take longer than the `:shutdown_grace_period` (by default
15 seconds) are brutally killed and left as orphans. The ids of jobs
left in an executing state are listed in the event's `orphaned` meta.

This also adds `queue:shutdown` logging to the default logger. Only
queues that shutdown with orphaned jobs are logged, which makes it
easier to detect orphaned jobs and which jobs were affected.

Inspired by #1076 from @axelson. Closes #1076.