Link queue supervisor and midwife for restarts
When a producer crashes it brings the queue's supervisor down with it.
With enough database errors, the producer may crash repeatedly enough to
exhaust restarts and bring down the DynamicSupervisor in charge of all
queues.

Now the supervisor is linked to the midwife to ensure that the midwife
restarts as well, and it restarts all of the queues.