Use concat operator for error concatenation
The standard `push` operation for updates is designed for arrays and
uses `array_append` internally. This replaces all use of `push` with a
fragment that uses the `||` operator instead, which works for both
arrays and jsonb.

CRDB (Cockroach) doesn't support arrays of jsonb, but they do support
simple jsonb columns. Now we can append to the errors column in either
format for CRDB compatibility.