Add rescue_jobs/3 callback and corresponding Dolphin implementation (#1199)
The Lifeline plugin formerly used two queries to rescue jobs—one to mark
jobs with remaining attempts as `available` and another that `discarded`
the remaining stuck jobs. Those are now combined into a single callback,
with the base definition in the `Basic` engine.

MySQL won't accept a select in an update statement. The dolphin
implementation of rescue_jobs/3 uses multiple queries to return the
relevant telemetry data and make multiple updates.