Delegate to JSON and fallback to Jason
Adds a new `Oban.JSON` module that detects whether the new JSON module
is available at compile time. If it isn't available, then it falls back
to Jason, and if Jason isn't available (which is extremely rare) then it
warns about a missing module.

This approach was chosen over a config option for backward compatibility
because Oban will only support the JSON module once the minimum
supported Elixir version is v1.18.

Closes #1213