improvement: make oban installer idempotent
by using `on_exists: :skip` when generating the migration, it is
safe for other igniter installers (like the one I'm working on)
to add `{:oban, "~> 2.0"}` to their `installs` list in their
igniters. If oban is present, the installer does nothing, but if
not the user gets a full oban installation.

Also removes an unnecessary `add_dep` call which prompts users
to change the dep in cases where they have something like
`~> 2.0` instead of exactly `~> 2.18`.