Fix :get_dynamic_repo with releases
Lave Ducia here, long-time user, first-time contributor!

It seems that the recommendation for a [separate Oban pool](https://hexdocs.pm/oban/scaling.html#pooling) does not work
with OTP releases, as anonymous functions cannot be used in
configuration.

This PR allows an MFA tuple to be passed instead, and updates the
documentation to recommend this instead.