Convert queue_time to native time unit as per docs (#1240)
The telemetry docs state that measurements are recorded in `native` time units. However, since 566f0f this has not been the case for `queue_time`.

It works anyway, because native and nanosecond is of the same resolution. Erlang gives no such guarantee though: https://www.erlang.org/doc/apps/erts/erlang#t:time_unit/0

Although potentially irrelevant, this ensures that if the resolution of the time unit changes in the future, the code will still work as expected.