fix: measurements from the Executor process, not parent
When Producer starts jobs, it calls `Executor.new` in its own process, before handing it off to `Task.Supervisor.async_nolink` to call `Executor.call`. That means we currently write in the Producer pid, and take measurements from that process.

This was the simplest approach I could think of to have the executor read its own measurements.