Set state to scheduled when updating timestamp
When the `scheduled_at` timestamp is set during `update_job`, the state
is automatically set to "scheduled". This mirrors the functionality of
`insert_job`.

Note that the value of the timestamp isn't considered. Setting a
timestamp in the past will set the job as "scheduled". This isn't a
problem in practice because the stager will change the state to
"available" on the next cycle anyhow.

Closes #1353