Order stage query to maximize compound index usage
The core compound index couldn't be utilized by staging queries when the
planner estimated a large number of hits. Changing the query to order by
`scheduled_at` and `id`, it becomes an index scan.

It would still be more efficient to use a dedicated index, but OSS
doesn't have any mechanisms for automatic concurrent index creation and
we have to save it for later.

Closes #1362